pyproject.toml
[tool.poetry]
name = "hannah"
version = "0.1.0"
description = "Hardware Accelerator and Neural Network Architecture seArcH"
authors = [
    "Christoph Gerum <christoph.gerum@uni-tuebingen.de>",
    "Adrian Frischknecht <adrian.frischknecht@uni-tuebingen.de>",
    "Felix Lorenz <felix.lorenz@student.uni-tuebingen.de>",
    "Tobias Hald <tobias.hald@student.uni-tuebingen.de>"
]
packages = [
    {include='hydra_plugins'}
]


[tool.poetry.dependencies]
python = ">3.8 <3.11"
pwlf = "^2.0.0"
gitdb2 = "2.0.6"
chainmap = "^1.0.3"
torch = "2.0.0"
pytorch-lightning = "^1.8.3.post1"
torchaudio = "^2.0.1"
torchvision = "^0.15.1"
SoundFile = "^0.10.3"
zipp = "^3.3.1"
bitstring = "^3.1.7"
wfdb = "^3.2.0"
lightning-bolts = "^0.3.2"
tqdm="^4.60.0"
seaborn="^0.11.2"
pandas="^1.2.4"
hydra-optuna-sweeper = {version = "^1.1.1"}

hydra-core = "^1.1.0"
hydra-joblib-launcher = ">1.1.0"
hydra-submitit-launcher = ">=1.2.0"

tabulate = "^0.8.3"
GitPython = "^3.1.14"
h5py = "^3.2.1"
sympy = "^1.7.1"
nvsmi = "^0.4.2"
libsvm="^3.23.0.4"
fairscale = "^0.4.5"
numpy = "^1.22.2 <1.24"
scikit-learn = "^1.1.1"
hydra-colorlog = "^1.2.0"
z3-solver = "^4.11.2.0"
xgboost = "^1.5.2"

# optional dependencies for onnx, and onnxtf backends
onnxruntime = {version = "^1.4.0", optional = true}
onnx-tf = {version = "^1.9.0", optional = true}

# optional dependencies for tvm backend
hannah-tvm = {path = "external/hannah-tvm", optional = true, extras = ["automate", "micro"], develop=true}
networkx = "^2.8.8"
dgl = "^0.9.1"

#dependencies for vision tasks
imagecorruptions = {version="^1.1.2", optional = true}
timm = {version = "^0.6.11", optional = true}
pycocotools = {version = "^2.0.6", optional = true}
gdown = {version = "^4.5.3", optional = true}
albumentations = {version = "^1.3.0", optional = true}
kornia = {version="^0.6.4", optional=true}
lightning = "^1.8.3.post1"
dvc = {version = "^2.43.1", optional = true}
dvclive = {version = "^1.3.4", optional = true}



[tool.poetry.dev-dependencies]
pytest = ">=7.2.0"
pre_commit = ">=2.7.1"
pytest-cov = ">=2.10.1"
rope = ">=0.19.0"
mypy = ">=0.910"
torch-tb-profiler = ">=0.2.0"
black = ">=22.6.0"
coverage = ">=6.3.2"
isort = ">=5.10.1"
pyment = ">=0.3.3"
mkdocstrings = {version = ">=0.19.0", extras = ["python"]}
mkdocs-gen-files = ">=0.3.4"
mkdocs-literate-nav = ">=0.4.1"
flaky = ">=3.7.0"


[tool.poetry.extras]
tf-backend = ["onnx-tf"]
onnxrt-backend = ["onnxruntime"]
tvm = ["hannah-tvm"]
vision = ["pycocotools", "albumentations", "imagecorruptions", "timm", "gdown", "kornia"]
dvc = ["dvc", "dvclive"]

[tool.poetry.scripts]
hannah-train = 'hannah.tools.train:main'
hannah-characterize = 'hannah.tools.characterize:main'
hannah-eval = 'hannah.tools.eval:main'
hannah-objectdetection-eval = 'hannah.tools.objectdetection_eval:main'
hannah-nas-eval = "hannah.nas.eval.__main__:main"

#[[tool.poetry.source]]
#name = "pytorch"
#url = "https://download.pytorch.org/whl/cu113"
#default = false
#secondary = true


# FIXME (gerum): enable dgl with cuda support again
#[[tool.poetry.source]]
#name = "dgl"
#url = "https://data.dgl.ai/wheels/repo.html"
#default = false
#secondary = false


[tool.poetry.group.dev.dependencies]
pyre-check = "^0.9.17"
pytest-xdist = "^3.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
